Manually stringify Payload#detail_call_info's contents
The `arguments` and `return_value` of a call info can sometimes be an
Array or a Hash. And both of their `#to_s` method are aliased to
`#inspect`. This means they'll always call `#inspect` to their elements
when we try to stringify them.

This behavior sometimes creates noises when they contain big objects,
like a Rack request. So we need to stringify  their elements manually to reduce the noise.