Add CollectionProxy class
As we're building more helpers, there are cases that we'll need to
return multiple device objects from one helper method. In those cases,
it'll be hard to do method chaining because the return value would be an
Array.

So this commit adds a class called CollectionProxy that holds multiple
device instances and passes chained method to them.