Add hijack_attr_methods option
Because TracePoint doesn't support tracking attribute reader/writer
methods (which is a bug), those method calls are currently undetectable.

However, in most cases, we should still be able to track them by overriding
the generated methods with normal method definition. But since this
mutates the object's interfaces directly, it can still cause unexpected
issues. So the behavior needs to be activated explicitly with the
`hijack_attr_methods` option.